ARM: dts: msm: make SDE HDMI node non-pluggable
Currently DRM userspace does not listen to HDMI events.
It does not process the HPD events to trigger the display
turn-on sequence.

Hence this causes a display timeout in the external display
driver and prevents audio initialization.

Making the SDE HDMI display non-pluggable till DRM userspace
support is present to handle HDMI events.

Change-Id: Ia8d96b585c9179ad2bf00b8e494db7f799984ba9
Signed-off-by: Abhinav Kumar <abhinavk@codeaurora.org>