ARM: dts: msm: make SDE HDMI node non-pluggable
Currently DRM userspace does not listen to HDMI events.
It does not process the HPD events to trigger the display
turn-on sequence.

Hence this causes a display timeout in the external display
driver and prevents audio initialization.

Making the SDE HDMI display non-pluggable till DRM userspace
support is present to handle HDMI events.

Change-Id: Ia037586d85d46400eeac8d5b39f6718f1e6709dc
Signed-off-by: Suresh Reddy Yellala <suresh@codeaurora.org>