ARM: dts: msm: designate LDO10A as a parent of LDO7B on sdm660
On SDM660, there is a HW bug that could lead to chip damage if
the 3.1V rail (normally provided by PM660 LDO7B) is turned on
while the 1.8V (LDO10A) rail is off. A simple workaround is to
ensure that LDO10A is turned on prior to LDO7B. Achieve this by
making pm660_l10 the parent-supply of the pm6601_l7 regulator.

Change-Id: I748d2b8526066e746002ad007789c10422c12c68
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>